Remove duplicated system module
