+++ executor/wasm_host/Cargo.toml
@@ -14,7 +14,7 @@ bytes = "1.10"
 casper-executor-wasm-common = { version = "0.1.3", path = "../wasm_common" }
 casper-executor-wasm-interface = { version = "0.1.3", path = "../wasm_interface" }
 casper-storage = { version = "2.1.1", path = "../../storage" }
-casper-types = { version = "6.0.1", path = "../../types" }
+casper-types = { version = "6.0.1", path = "../../types", features = ["std", "testing"] }
 either = "1.15"
 num-derive = { workspace = true }
 num-traits = { workspace = true }
