+++ executor/wasm_host/src/lib.rs
@@ -2,4 +2,4 @@
 pub(crate) mod abi;
 pub mod context;
 pub mod host;
-pub(crate) mod system;
+pub mod system;
